更新: 软光栅L6/完成重构, 实现了Gouraud着色器

+++ Rasterizer/core/include/gl_soft.cpp
@@ -1,13 +1,13 @@
 ﻿#include "gl_soft.h"
 
-Vec3f barycentric(Vec3i A, Vec3i B, Vec3i C, Vec3i P) {
+Vec3f barycentric(Vec3f A, Vec3f B, Vec3f C, Vec3f P) {
 	// 计算三角形ABC中点P对应的重心坐标
 	// 任意两边的向量
 	Vec3f s[2];
 	for (int i = 2; i--; ) {
-		s[i][0] = C[i] - A[i];
-		s[i][1] = B[i] - A[i];
-		s[i][2] = A[i] - P[i];
+		s[i][0] = int(C[i]) - int(A[i]);
+		s[i][1] = int(B[i]) - int(A[i]);
+		s[i][2] = int(A[i]) - int(P[i]);
 	}
 	// 叉乘得到的第三轴调整后就是重心坐标
 	Vec3f u = cross(s[0], s[1]);
