+++ Rasterizer/core/include/gl_soft.h
@@ -26,7 +26,7 @@ public:
 
 //void triangle(Vec3f* pts, int** zbuffer, TGAImage& image, f_Shader* f_shader);
 
-Vec3f barycentric(Vec3i A, Vec3i B, Vec3i C, Vec3i P);
+Vec3f barycentric(Vec3f A, Vec3f B, Vec3f C, Vec3f P);
 
 Matrix makeModelRotationMat(Vec3f xyz);
